drop global flags:
As nx is becoming more generic, we cannot assume arg number is going to be small
enough to scan through all of it, looking for specific flags.
The scan will become horribly slow if a glob spawns thousands of positionals.

Also having explicitly declared and handled flags, eventhough it adds repetition
in some cases, allows to have explicit options and behaviour.